+++ src/Migration.php
@@ -93,7 +93,7 @@ final class Migration
     {
         $installedVersion = get_option(JOOOSI_FON::WP_OPTION . '_version', false);
 
-        if (! is_string($installedVersion) || version_compare($installedVersion, JOOOSI_FON::VERSION, '>=')) {
+        if (is_string($installedVersion) && version_compare($installedVersion, JOOOSI_FON::VERSION, '>=')) {
             return;
         }